chore: move the snapshot machinery into its own module
